+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Added a new constructor for `TransactionExecutor` that accepts `ExecutionOptions` (#1502).
 - [BREAKING] Introduce errors in `MockChain` API (#1508).
 - [BREAKING] `TransactionAdviceInputs` cannot return `Err` anymore (#1517).
+- Implement serialization for `LexicographicWord` (#1524).
 
 ## 0.9.5 (2025-06-20) - `miden-lib` crate only